- Smoke Test Seg Fault fix
(cherry picked from commit 133ad61365ff12fea97430e305e33c83d8793723)
