+++ geonode/base/populate_test_data.py
@@ -28,47 +28,16 @@ from django.core.files.uploadedfile import SimpleUploadedFile
 from django.utils import timezone
 from geonode.layers.models import Layer
 from geonode.base.models import TopicCategory
-from geonode.maps.models import Map, MapLayer
+from geonode.maps.models import Map
 from geonode.documents.models import Document
 from geonode.people.models import Profile
 from geonode import geoserver, qgis_server  # noqa
-from geonode.utils import check_ogc_backend
 from itertools import cycle
 from taggit.models import Tag
 from taggit.models import TaggedItem
 from uuid import uuid4
 import os.path
 import six
-
-
-def disconnect_signals():
-    """Disconnect signals for test class purposes."""
-    from django.db.models import signals
-    from geonode.geoserver.signals import geoserver_pre_save_maplayer
-    from geonode.geoserver.signals import geoserver_post_save_map
-    from geonode.geoserver.signals import geoserver_pre_save
-    from geonode.geoserver.signals import geoserver_post_save
-    signals.pre_save.disconnect(geoserver_pre_save_maplayer, sender=MapLayer)
-    signals.post_save.disconnect(geoserver_post_save_map, sender=Map)
-    signals.pre_save.disconnect(geoserver_pre_save, sender=Layer)
-    signals.post_save.disconnect(geoserver_post_save, sender=Layer)
-
-
-def reconnect_signals():
-    """Reconnect signals for test class purposes."""
-    from django.db.models import signals
-    from geonode.geoserver.signals import geoserver_pre_save_maplayer
-    from geonode.geoserver.signals import geoserver_post_save_map
-    from geonode.geoserver.signals import geoserver_pre_save
-    from geonode.geoserver.signals import geoserver_post_save
-    signals.pre_save.connect(geoserver_pre_save_maplayer, sender=MapLayer)
-    signals.post_save.connect(geoserver_post_save_map, sender=Map)
-    signals.pre_save.connect(geoserver_pre_save, sender=Layer)
-    signals.post_save.connect(geoserver_post_save, sender=Layer)
-
-
-if check_ogc_backend(geoserver.BACKEND_PACKAGE):
-    disconnect_signals()
 
 # This is used to populate the database with the search fixture data. This is
 # primarily used as a first step to generate the json data for the fixture using
